refactor: Extract command handlers from conversation.rs (1029->367 LOC)
Split conversation.rs into two focused modules:
- conversation.rs: message processing, streaming, and tool-loop orchestration
- command_handlers.rs: slash command dispatch and all handler functions

The handle_command function is re-exported from conversation.rs so that
interactive.rs continues to compile with no import changes.
